Leveling System / Movement / Doxygen Format
Implementation of Experience manager that holds a players level, current experience and does internal calculations of the leveling formula. Invokes an event when a player levels up so that we can later assign listeners. Slight changes to movement for slight Air control allowing Axes input when the player is in the air. Doxygen Comment format implementation on all scripts for Automated Code Documentation